fixed double comments issue, now the comments are fetched once in index.ts and loaded depending on the product detail
