+++ src/pages/ProductDetailsPage.ts
@@ -3,8 +3,6 @@ import routeParams from "../routeParams";
 import { store, catalog } from '../index';
 import productDetails from '../views/productDetails';
 import productDetailsComponent from "../components/productDetailsComponent";
-import CommentInterface from '../interfaces/CommentInterface';
-import selectComments from '../helpers/selectComments';
 import commentsList from '../views/commentsList';
 import displayHamburgerMenu from '../helpers/hamburger';
 import backToTop from '../helpers/backToTop';
@@ -52,29 +50,10 @@ function renderDetail(): void {
 
 function renderComments() {
 
-    async function getCommentsFromApi() {
+    commentsList(store.getComments());
 
-        try {
-
-          let res = await fetch('https://jsonplaceholder.typicode.com/comments'),
-          json: CommentInterface[] = await res.json();
-    
-          if (!res.ok) { throw new Error("algo salió mal")}
-    
-          selectComments(store, json);
-    
-          commentsList(store.getComments());
-    
-        }
-        catch(err) {
-
-            console.log(err)
-
-        }
-
-    }
-    getCommentsFromApi();
 }
+
 
 function toggleCommentsOnClick() {
